change text_size to float datatype

+++ models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, ForeignKey, Text, Boolean
+from sqlalchemy import Column, Integer, String, ForeignKey, Text, Boolean,Float
 from sqlalchemy.dialects.mysql import JSON
 from sqlalchemy.orm import relationship
 from sqlalchemy.ext.declarative import declarative_base
@@ -33,7 +33,7 @@ class Cv(Base):
     is_skills = Column(Boolean, default=True)
     is_loisirs = Column(Boolean, default=True)
     is_active = Column(Boolean,default=True)
-    text_size = Column(Integer)
+    text_size = Column(Float)
     order_cate = Column(String(256))
     user_id = Column(Integer, ForeignKey('users.id', ondelete='CASCADE'))
     user = relationship("User", back_populates="cvs")
